% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addParameters.R
\name{add_params}
\alias{add_params}
\title{Add parameters to bench}
\usage{
add_params(b, label, ..., .list = NULL, overwrite = FALSE)
}
\arguments{
\item{b}{bench object}

\item{label}{name for method, must already exist}

\item{...}{named parameter, value pairs to be used with func}

\item{.post}{}
}
\value{
a bench
}
\description{
Add parameters to bench
}
\author{
Patrick Kimes
}
