% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefaultMetrics.R
\name{availableMetrics}
\alias{availableMetrics}
\alias{addDefaultMetrics}
\title{Add default performance metrics to a \code{\link{SummarizedBenchmark}} object.}
\usage{
availableMetrics()

addDefaultMetrics(object, metrics = NULL)
}
\arguments{
\item{object}{A \code{\link{SummarizedBenchmark}} object.}

\item{metrics}{A character vector with performance metric names to be added.
See function \code{availableMetrics}.}
}
\value{
A \code{\link{SummarizedBenchmark}} object with performance functions.
}
\description{
List default performance metrics available in this package.

This function adds predefined performance metrics to a \code{\link{SummarizedBenchmark}}
object, according to the assay name. For example, if an assay is named "qvalue",
it will add a the performance metric "rejections".
}
\details{
This function adds a set of predefined performance metrics according
to the names of the assays. Specifically, if an assay is called "qvalue" it will
add the "rejections" metric by default. Furthermore, if the qvalue assay has
information about the ground truths in the rowData, the metrics TPR (true
positive rate), "TNR" (true negative rate), "FPR" (false positive rate) and
"FNR" (false negative rate) will be added. See function \code{availableMetrics()}
to retrieve a complete list of metrics defined in this package.
}
\examples{

data( sb )
sb <- addDefaultMetrics( sb )

}
\author{
Alejandro Reyes
}
