% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMethod.R
\name{addMethod}
\alias{addMethod}
\title{Add a new method to a BenchDesign}
\usage{
addMethod(b, blabel, bfunc, bpost = NULL, ...)
}
\arguments{
\item{b}{BenchDesign object}

\item{blabel}{name for the method}

\item{bfunc}{function to be benchmarked}

\item{bpost}{optional post-processing function that takes output of func,
e.g. a getter method, ignored if NULL (default = NULL)}

\item{...}{named parameter, value pairs to be used with func}
}
\description{
Add a new method to a BenchDesign
}
\author{
Patrick Kimes
}
