% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformanceMetrics.R
\name{addPerformanceFunction}
\alias{addPerformanceFunction}
\title{Add a performance metric definition to a \code{\link{SummarizedBenchmark}} object.}
\usage{
addPerformanceFunction(object, evalMetric, assay, evalFunction)
}
\arguments{
\item{object}{A \code{\link{SummarizedBenchmark}} object.}

\item{evalMetric}{A string with the name of the evaluation metric.}

\item{assay}{A string with an assay name. Indicates the assay that should be
given as input to this performance metric.}

\item{evalFunction}{A function that calculates a performance metric. It should
contain at least two arguments, query and truth, where query is the output vector
of a method and truth is the vector of true values. If additional parameters
are specified, they must contain default values.}
}
\value{
A \code{\link{SummarizedBenchmark}} object.
}
\description{
This is a function to define performance metrics for benchmarking methods.
The function is saved into the \code{performanceFunctions} slot.
}
\examples{

data( sb )
sb <- addPerformanceFunction(
   object=sb,
   assay="qvalue",
   evalMetric="TPR",
   evalFunction = function( query, truth, alpha=0.1 ){
       goodHits <- sum( (query < alpha) & truth == 1 )
       goodHits / sum(truth == 1)
   }
)

}
