% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformanceMetrics.R
\name{addPerformanceMetric}
\alias{addPerformanceMetric}
\title{Add performance metric to SummarizedBenchmark object}
\usage{
addPerformanceMetric(object, evalMetric, assay, evalFunction = NULL)
}
\arguments{
\item{object}{A \code{\link[=SummarizedBenchmark-class]{SummarizedBenchmark}} object.}

\item{evalMetric}{A string with the name of the evaluation metric.}

\item{assay}{A string with an assay name. Indicates the assay that should be
given as input to this performance metric.}

\item{evalFunction}{A function that calculates a performance metric. It should
contain at least two arguments, query and truth, where query is the output vector
of a method and truth is the vector of true values. If additional parameters
are specified, they must contain default values. If NULL, the 'evalMetric' string must
be the name of a predefined metric available through 'availableMetrics()$function'.}
}
\value{
A \code{\link[=SummarizedBenchmark-class]{SummarizedBenchmark}} object.
}
\description{
This is a function to define performance metrics for benchmarking methods.
The function is saved into the \code{performanceMetrics} slot.
}
\examples{
data( sb )
sb <- addPerformanceMetric(
   object=sb,
   assay="qvalue",
   evalMetric="TPR",
   evalFunction = function( query, truth, alpha=0.1 ){
       goodHits <- sum( (query < alpha) & truth == 1 )
       goodHits / sum(truth == 1)
   }
)

}
\seealso{
\code{\link{availableMetrics}}, \code{\link{performanceMetrics}}
}
\author{
Alejandro Reyes
}
