% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-soneson2016.R
\docType{data}
\name{allSB}
\alias{allSB}
\alias{allSB,quantSB}
\title{SummarizedBenchmark object of isoform quantification results}
\source{
Mouse BodyMap (Li et al, 2014). SRA accession numbers SRR5273705, SRR5273689, SRR5273699 and SRR5273683.
}
\description{
This object is a \code{SummarizedBenchmark} object containing isoform quantifications from salmon, sailfish and kallisto
from 4 mouse samples (2 hearts and 2 brains) part of the Mouse BodyMap. Its generation is described in one of the vignettes of this package.
}
\examples{
data( quantSB )
}
\keyword{data}
