% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedBenckmark-accessors.R
\docType{methods}
\name{assayNames-setter}
\alias{assayNames-setter}
\title{assayNames setter}
\usage{
\S4method{assayNames}{SummarizedBenchmark,character}(x, ...) <- value
}
\arguments{
\item{x}{A \code{SummarizedBenchmark} object.}

\item{...}{Futher arguments, perhaps used by methods}

\item{value}{A character vector.}
}
\value{
Modfied \code{SummarizedBenchmark} object.
}
\description{
assayNames setter
}
\examples{

data( sb )
assayNames( sb )[2] <- "log2FC"

}
\author{
Alejandro Reyes
}
