% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefaultMetrics.R
\name{availableMetrics}
\alias{availableMetrics}
\title{availableMetrics}
\usage{
availableMetrics()
}
\value{
A data.frame with available performance metrics.
}
\description{
List default performance metrics available in this package.
}
\examples{
availableMetrics()

}
