% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BDData-class}
\alias{BDData-class}
\title{BDData class}
\description{
Container for data in a BenchDesign object.
}
\section{Slots}{

\describe{
\item{\code{data}}{a list or MD5 hash of the data.}

\item{\code{type}}{a character string indicating whether the data slot
contains the 'data' or a 'md5hash' of the data.}
}}

\seealso{
\code{\link{BDData}}, \code{\link{BenchDesign-class}}, \code{\link{BDMethod-class}}, \code{\link{BDMethodList-class}}
}
\author{
Patrick Kimes
}
