% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDData-show.R
\docType{methods}
\name{BDData-show}
\alias{BDData-show}
\title{Show BDData object}
\usage{
\S4method{show}{BDData}(object)
}
\arguments{
\item{object}{BDData object to show}
}
\description{
Show BDData object
}
