% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchDesign-comparisons.R
\name{BDDataHash}
\alias{BDDataHash}
\title{Convert BDData to BDData with MD5 Hash}
\usage{
BDDataHash(bdd)
}
\arguments{
\item{bdd}{a \code{BDData} object}
}
\value{
\code{BDData} object with type = "md5hash".
}
\description{
Simple converting function to replace data object in
BDData object with MD5 hash value.
}
\author{
Patrick Kimes
}
