% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BDMethodList-accessors.R,
%   R/BenchDesign-accessors.R
\docType{methods}
\name{BDMethod<-}
\alias{BDMethod<-}
\alias{BDMethod<-,BDMethodList,character,BDMethod-method}
\alias{BDMethod<-,BDMethodList,character,NULL-method}
\alias{BDMethod<-,BenchDesign,character,BDMethod-method}
\alias{BDMethod<-,BenchDesign,character,NULL-method}
\title{Set method in list or BenchDesign object}
\usage{
BDMethod(x, i) <- value

\S4method{BDMethod}{BDMethodList,character,BDMethod}(x, i) <- value

\S4method{BDMethod}{BDMethodList,character,`NULL`}(x, i) <- value

\S4method{BDMethod}{BenchDesign,character,BDMethod}(x, i) <- value

\S4method{BDMethod}{BenchDesign,character,`NULL`}(x, i) <- value
}
\arguments{
\item{x}{\code{\link[=BenchDesign-class]{BenchDesign}} or
\code{\link[=BDMethodList-class]{BDMethodList}} object.}

\item{i}{character name of method.}

\item{value}{\code{\link[=BDMethod-class]{BDMethod}} or \code{NULL}.}
}
\value{
modified BenchDesign object
}
\description{
Adds, replaces or removes a named \code{\link[=BDMethod-class]{BDMethod}} method
in a \code{\link[=BDMethodList-class]{BDMethodList}} or
\code{\link[=BenchDesign-class]{BenchDesign}} object with a specified
\code{\link[=BDMethod-class]{BDMethod}} object.

An existing method can be removed by setting the value to \code{NULL}.
}
\seealso{
\code{\link{BDMethod}}
}
\author{
Patrick Kimes
}
