% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDMethod-show.R
\docType{methods}
\name{BDMethod-show}
\alias{BDMethod-show}
\title{Show BDMethod object}
\usage{
\S4method{show}{BDMethod}(object)
}
\arguments{
\item{object}{BDMethod object to show}
}
\description{
Show BDMethod object
}
