% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BDMethod-constructor.R,
%   R/BDMethodList-accessors.R, R/BenchDesign-accessors.R
\docType{methods}
\name{BDMethod}
\alias{BDMethod}
\alias{BDMethod,quosure-method}
\alias{BDMethod,function-method}
\alias{BDMethod,BDMethodList-method}
\alias{BDMethod,BenchDesign-method}
\title{Create a new BDMethod object}
\usage{
BDMethod(x, params = rlang::quos(), post = NULL, meta = NULL, ...)

\S4method{BDMethod}{quosure}(x, params = rlang::quos(), post = NULL,
  meta = NULL, ...)

\S4method{BDMethod}{`function`}(x, params = rlang::quos(), post = NULL,
  meta = NULL, ...)

\S4method{BDMethod}{BDMethodList}(x, i = 1)

\S4method{BDMethod}{BenchDesign}(x, i = 1)
}
\arguments{
\item{x}{main method function or function quosure. Alternative, may be a
BDMethodList or BenchDesign object from which the BDMethod should be
extracted.}

\item{params}{list of quosures specifying function parameters.
(default = \code{rlang::quos()})}

\item{post}{list of functions to be applied to the output of \code{x}.
(default = NULL)}

\item{meta}{list of metadata. (default = NULL)}

\item{...}{other parameters.}

\item{i}{integer index or character name of BDMethod in BDMethodList or
BenchDesign object.}
}
\value{
BDMethod object
}
\description{
Initializes a new BenchDesign method (BDMethod) object for benchmarking methods.
}
\examples{
BDMethod(x = base::mean)

BDMethod(x = function(x) { x^2 }, post = base::sqrt,
         meta = list(note = "simple example"))

bd <- BenchDesign(method1 = BDMethod(x = base::mean))
BDMethod(bd, "method1")

}
