% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BDMethodList-accessors.R
\docType{methods}
\name{BDMethodList-accessors}
\alias{BDMethodList-accessors}
\alias{BDMethod<-,BDMethodList,character,BDMethod-method}
\alias{BDMethod<-,BDMethodList,character,NULL-method}
\alias{BDMethod,BDMethodList-method}
\title{BDMethodList Accessors}
\usage{
\S4method{BDMethod}{BDMethodList,character,BDMethod}(x, i) <- value

\S4method{BDMethod}{BDMethodList,character,`NULL`}(x, i) <- value

\S4method{BDMethod}{BDMethodList}(x, i = 1)
}
\arguments{
\item{x}{BDMethodList object.}

\item{i}{character name or integer index of a BDMethod object.
(Note: must be a character name for replacement operations.)}

\item{value}{replacement object, either BDMethod or NULL object.}
}
\value{
modified BDMethodList object
}
\description{
These methods can be used to access, set, and remove elements of
a BDMethodList object.
}
\author{
Patrick Kimes
}
