% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/BDMethodList-constructor.R
\docType{class}
\name{BDMethodList-class}
\alias{BDMethodList-class}
\alias{BDMethodList}
\alias{BDMethodList,ANY-method}
\alias{coerce}
\title{BDMethodList class}
\usage{
BDMethodList(..., object = NULL)

\S4method{BDMethodList}{ANY}(..., object = NULL)
}
\arguments{
\item{...}{a named list of BDMethod objects}
}
\value{
BDMethodList object
}
\description{
Extension of the SimpleList class to contain a list of BDMethod
objects.

Initialized a new SimpleList of BenchDesign method (BDMethod) objects.
}
\examples{
BDMethodList()

}
\author{
Patrick Kimes
}
