% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BenchDesign-accessors.R
\docType{methods}
\name{BDMethodList<-}
\alias{BDMethodList<-}
\alias{BDMethodList<-,BenchDesign,BDMethodList-method}
\title{Set method list in BenchDesign object}
\usage{
BDMethodList(x) <- value

\S4method{BDMethodList}{BenchDesign,BDMethodList}(x) <- value
}
\arguments{
\item{x}{BenchDesign object.}

\item{value}{\code{\link[=BDMethodList-class]{BDMethodList}} list of methods.}
}
\value{
modified BenchDesign object
}
\description{
Replaces the \code{\link[=BDMethodList-class]{BDMethodList}} list of methods
in a \code{\link[=BenchDesign-class]{BenchDesign}} object with a specified
\code{\link[=BDMethodList-class]{BDMethodList}} object.
}
\seealso{
\code{\link{BDMethod}}, \code{\link{BDMethodList}}
}
\author{
Patrick Kimes
}
