% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDMethodList-show.R
\docType{methods}
\name{show,BDMethodList-method}
\alias{show,BDMethodList-method}
\title{Show BDMethodList object}
\usage{
\S4method{show}{BDMethodList}(object)
}
\arguments{
\item{object}{BDMethodList object to show}
}
\description{
Show BDMethodList object
}
