% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BDMethodList-constructor.R,
%   R/BenchDesign-accessors.R, R/SummarizedBenckmark-accessors.R
\docType{methods}
\name{BDMethodList}
\alias{BDMethodList}
\alias{BDMethodList,ANY-method}
\alias{coerce}
\alias{BDMethodList,BenchDesign-method}
\alias{BDMethodList,SummarizedBenchmark-method}
\title{Create a new BDMethodList object}
\usage{
BDMethodList(..., x = NULL)

\S4method{BDMethodList}{ANY}(..., x = NULL)

\S4method{BDMethodList}{BenchDesign}(..., x = NULL)

\S4method{BDMethodList}{SummarizedBenchmark}(..., x = NULL)
}
\arguments{
\item{...}{a named list of BDMethod objects}

\item{x}{a BenchDesign or SummarizedBenchmark object to extract
the BDMethodList from. (default = NULL)}
}
\value{
BDMethodList object
}
\description{
Initialized a new SimpleList of BenchDesign method (BDMethod) objects.
}
\examples{
BDMethodList()

}
\author{
Patrick Kimes
}
