% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BenchDesign-accessors.R
\docType{methods}
\name{BDData<-}
\alias{BDData<-}
\alias{BDMethod<-}
\alias{BDMethodList<-}
\alias{BenchDesign-accessors}
\alias{BDData<-,BenchDesign,BDDataOrNULL-method}
\alias{BDMethod<-,BenchDesign,character,BDMethod-method}
\alias{BDMethod<-,BenchDesign,character,NULL-method}
\alias{BDMethodList<-,BenchDesign,BDMethodList-method}
\alias{BDMethod,BenchDesign-method}
\alias{BDMethodList,BenchDesign-method}
\title{BenchDesign Accessors}
\usage{
BDData(x) <- value

BDMethod(x, i) <- value

BDMethodList(x) <- value

\S4method{BDData}{BenchDesign,BDDataOrNULL}(x) <- value

\S4method{BDMethod}{BenchDesign,character,BDMethod}(x, i) <- value

\S4method{BDMethod}{BenchDesign,character,`NULL`}(x, i) <- value

\S4method{BDMethodList}{BenchDesign,BDMethodList}(x) <- value

\S4method{BDMethod}{BenchDesign}(x, i = 1)

\S4method{BDMethodList}{BenchDesign}(..., x = NULL)
}
\arguments{
\item{x}{BenchDesign object.}

\item{value}{replacement object, either BDData, BDMethod, BDMethodList
or NULL object.}

\item{i}{character name or integer index of a BDMethod object.
(Note: must be a character name for replacement operations.)}
}
\value{
modified BenchDesign object
}
\description{
These methods can be used to access, set, and remove elements of
a BenchDesign object.
}
\author{
Patrick Kimes
}
