% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BenchDesign-class}
\alias{BenchDesign-class}
\title{BenchDesign class}
\description{
Along with the SummarizedBenchmark class, one of the two main
classses of the SummarizedBenchmark package. The BenchDesign class serves as
a container for both the set of methods to be benchmarked and optionally the
data to be used for benchmarking.

Methods are organized as BDMethod objects and stored in as a list using the
BDMethodList class. The BDData class is used to store benchmarking data, or
in some cases, just the MD5 hash of the original data set. Any list object,
including data.frame objects, can be specified for data. More details on the
component classes are provided in the corresponding class documentation.

For details on how to create new BenchDesign objects, see the
documentation for the \code{\link{BenchDesign}} constructor.
}
\section{Slots}{

\describe{
\item{\code{data}}{a list containing the data to be used in the benchmark.}

\item{\code{methods}}{a \code{BDMethodList} list of \code{BDMethod} objects to be
compared in the benchmark.}
}}

\seealso{
\code{\link{BenchDesign}}, \code{\link{BDMethod-class}}, \code{\link{BDMethodList-class}}, \code{\link{BDData-class}}
}
\author{
Patrick Kimes
}
