% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchDesign-show.R
\docType{methods}
\name{BenchDesign-show}
\alias{BenchDesign-show}
\title{Show BenchDesign object}
\usage{
\S4method{show}{BenchDesign}(object)
}
\arguments{
\item{object}{BenchDesign object to show}
}
\description{
Show BenchDesign object
}
