% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildBench.R
\name{buildBench}
\alias{buildBench}
\title{Make SummarizedBenchmark from BenchDesign}
\usage{
buildBench(b, data = NULL, truthCol = NULL, ptabular = TRUE)
}
\arguments{
\item{b}{BenchDesign object.}

\item{data}{Data set to be used for benchmarking, will take priority over
data set originally specified to BenchDesign object.
Ignored if NULL. (default = NULL)}

\item{truthCol}{Character name of column in data set corresponding to ground
truth values. If specified, column will be added to \code{groundTruth}
DataFrame for the returned SummarizedBenchmark object, and same name
will be used for assay. (default = NULL)}

\item{ptabular}{Whether to return method parameters with each parameter in a
separate column of the \code{colData} for the returned SummarizedBenchmark
object, i.e. in tabular form. If FALSE, method parameters are returned
as a single column with comma-delimited "key=value" pairs. (default = TRUE)}
}
\value{
\code{SummarizedBenchmark} object with single assay
}
\description{
Function to evaluate BenchDesign methods on supplied
data set to generate a SummarizedBenchmark.
}
\author{
Patrick Kimes
}
