% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildBench.R
\name{clean_methods}
\alias{clean_methods}
\title{helper function to convert method info to character for colData}
\usage{
clean_methods(b)
}
\description{
helper function to convert method info to character for colData
}
