% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchDesign-comparisons.R
\name{compareBDMethods}
\alias{compareBDMethods}
\title{Simple comparison of two BDMethod objects based on
meta data.}
\usage{
compareBDMethods(bdm1, bdm2)
}
\arguments{
\item{bdm1}{a \code{BDMethod} object}

\item{bdm2}{a \code{BDMethod} object}
}
\value{
logical value indicating whether the two objects produced the same
meta data.
}
\description{
Simple comparison of two BDMethod objects based on
meta data.
}
\author{
Patrick Kimes
}
