% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchDesign-comparisons.R
\name{compareBenchDesigns}
\alias{compareBenchDesigns}
\title{Compare BenchDesign objects}
\usage{
compareBenchDesigns(bd1, bd2)
}
\arguments{
\item{bd1}{a \code{BenchDesign} object}

\item{bd2}{a \code{BenchDesign} object}
}
\value{
logical value indicating whether the two objects have
methods producing the same meta data and data with the
same MD5 hashes.
}
\description{
Simple comparison of two BenchDesign objects based on
methods meta data and data MD5 hashs.
}
\author{
Patrick Kimes
}
