% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropBMethod.R
\name{dropBMethod}
\alias{dropBMethod}
\title{Remove method from BenchDesign object}
\usage{
dropBMethod(b, blabel)
}
\arguments{
\item{b}{BenchDesign object.}

\item{blabel}{Character name of method to be modified.}
}
\value{
Modified BenchDesign object.
}
\description{
This function takes a BenchDesign object and the name of a method
already defined in the object, and returns a reduced BenchDesign
object with the specified method removed.
}
\author{
Patrick Kimes
}
