% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropMethod.R
\name{dropMethod}
\alias{dropMethod}
\title{Remove method from BenchDesign object}
\usage{
dropMethod(bd, label)
}
\arguments{
\item{bd}{\code{\link[=BenchDesign-class]{BenchDesign}} object.}

\item{label}{Character name of method.}
}
\value{
Modified \code{\link[=BenchDesign-class]{BenchDesign}} object with
specified method dropped.
}
\description{
Takes a \code{\link[=BenchDesign-class]{BenchDesign}} object and the name of an
existing method
and returns a reduced \code{\link[=BenchDesign-class]{BenchDesign}} object with the
method removed.
}
\examples{
## empty BenchDesign
bench <- BenchDesign()

## add methods
bench <- addMethod(bench, label = "bonf", func = p.adjust,
                   params = rlang::quos(p = pval, method = "bonferroni"))
bench <- addMethod(bench, label = "BH", func = p.adjust,
                   params = rlang::quos(p = pval, method = "BH"))
BDMethodList(bench)

## remove methods
bench <- dropMethod(bench, label = "bonf")
BDMethodList(bench)

}
\seealso{
\code{\link{modifyMethod}}, \code{\link{expandMethod}}, \code{\link{addMethod}}
}
\author{
Patrick Kimes
}
