% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandBMethod.R
\name{expandBMethod}
\alias{expandBMethod}
\title{Expand method in BenchDesign object across parameter settings}
\usage{
expandBMethod(b, blabel, param = NULL, ..., .replace = FALSE,
  .overwrite = FALSE)
}
\arguments{
\item{b}{BenchDesign object.}

\item{blabel}{Character name of method to be modified.}

\item{param}{Character name of parameter to be modified. Should only be
specified if only one parameter should be replaced in original
method definition. Only one of `param` or `list` should be
specified. (default = NULL)}

\item{...}{If `param != NULL`, named values to use for
overwriting the specified parameter in the original method definition.
If `param = NULL`, then named lists of parameter sets to be used in
place of parameters in the original method definition. Each parameter set
should also be a named list of parameter, value pairs.
Names will be used as the new method names in the BenchDesign object.
An error will be returned if an existing method name is used.
(defualt = NULL)}

\item{.replace}{Logical whether original `blabel` object should be removed
if method expansion is successful. (default = FALSE)}

\item{.overwrite}{Logical whether to overwrite the existing list of
parameters (TRUE) or to simply add the new parameters to the existing
list (FALSE). (default = FALSE)}
}
\value{
Modified BenchDesign object.
}
\description{
This function takes a BenchDesign object and the name of a method
already defined in the object, and returns a modified BenchDesign
object with multiple copies of the method differing by the specified
parameter sets.
}
\author{
Patrick Kimes
}
