% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BDData-constructor.R
\docType{methods}
\name{hashBDData}
\alias{hashBDData}
\alias{hashBDData,BDData-method}
\alias{hashBDData,BenchDesign-method}
\title{Hash data in BDData object}
\usage{
hashBDData(object)

\S4method{hashBDData}{BDData}(object)

\S4method{hashBDData}{BenchDesign}(object)
}
\arguments{
\item{object}{a \code{BDData} or \code{BenchDesign} object}
}
\value{
an object of the same class as \code{object} with data
converted to a MD5 hash.
}
\description{
Repliaces data stored in a \code{\link[=BDData-class]{BDData}} object
with the MD5 hash of the data. If the data was already a MD5 hash, the
original object is returned unchanged. The method can be called directly
on \code{\link[=BenchDesign-class]{BenchDesign}} objects to hash the
underlying data as well.
}
\author{
Patrick Kimes
}
