% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchDesign.R
\name{is.BenchDesign}
\alias{is.BenchDesign}
\title{Reports whether x is a BenchDesign object}
\usage{
is.BenchDesign(x)
}
\arguments{
\item{x}{object to test}
}
\description{
Reports whether x is a BenchDesign object
}
\keyword{internal}
