% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedBenckmark-accessors.R
\docType{methods}
\name{mcols<-,SummarizedBenchmark-method}
\alias{mcols<-,SummarizedBenchmark-method}
\title{Set meta data columns in SummarizedBenchmark object}
\usage{
\S4method{mcols}{SummarizedBenchmark}(x, ...) <- value
}
\arguments{
\item{x}{A \code{\link[=SummarizedBenchmark-class]{SummarizedBenchmark}} object.}

\item{...}{Futher arguments, perhaps used by methods.}

\item{value}{A DataFrame of meta data.}
}
\value{
Modfied \code{\link[=SummarizedBenchmark-class]{SummarizedBenchmark}} object.
}
\description{
Modifies the \code{mcols} slot of a \code{\link[=SummarizedBenchmark-class]{SummarizedBenchmark}}
object.
}
\seealso{
\code{\link[=SummarizedBenchmark-class]{SummarizedBenchmark}}
}
\author{
Alejandro Reyes
}
