% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifyBMethod.R
\name{modifyBMethod}
\alias{modifyBMethod}
\title{Modify definition of method in BenchDesign object}
\usage{
modifyBMethod(b, blabel, ..., .overwrite = FALSE)
}
\arguments{
\item{b}{BenchDesign object.}

\item{blabel}{Character name of method to be modified.}

\item{...}{Named parameter, value pairs to overwrite
in method definition. This can include \code{bfunc},
\code{bpost}, and \code{bmeta} parameters. All other named parameters
will be added to the list of parameters to be passed to
\code{bfunc}.}

\item{.overwrite}{Logical whether to overwrite the existing list of
parameters to be passed to \code{bfunc} (TRUE), or to simply add
the new parameters to the existing list (FALSE).
(default = FALSE)}
}
\value{
Modified BenchDesign object.
}
\description{
This function takes a BenchDesign object and the name of a method
already defined in the object, and returns a modified BenchDesign
object with the specified changes made only to the named method.
At a minimum, a string name for the method, \code{blabel}, must be
specified in addition to the primary BenchDesign object.
}
\examples{
## with toy data.frame
df <- data.frame(pval = runif(100))
bd <- BenchDesign(df)

## add method
bd <- addBMethod(bd, blabel = "qv",
                 bfunc = qvalue::qvalue,
                 bpost = function(x) { x$qvalue },
                 bmeta = list(note = "storey's q-value"),
                 p = pval)

## modify method 'bmeta' property of 'qv' method
bd <- modifyBMethod(bd, blabel = "qv",
                    bmeta = list(note = "Storey's q-value"))

## verify that method has been updated
showBMethod(bd, "qv")

}
\author{
Patrick Kimes
}
