% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifyMethod.R
\name{modifyMethod}
\alias{modifyMethod}
\title{Modify method in BenchDesign object}
\usage{
modifyMethod(bd, label, params, .overwrite = FALSE)
}
\arguments{
\item{bd}{\code{\link[=BenchDesign-class]{BenchDesign}} object.}

\item{label}{Character name of method to be modified.}

\item{params}{Named quosure list created using \code{\link[rlang]{quos}} of
\code{parameter = value} paiars to replace in the method definition.
The \code{post}, and \code{meta} parameters of the method can be
modified using the special keywords, \code{bd.post}, and \code{bd.meta}
(the prefix denoting that these values should modify \code{\link[=BenchDesign-class]{BenchDesign}}
parameters). All other named parameters will be added to the list of
parameters to be passed to \code{func}.}

\item{.overwrite}{Logical whether to overwrite the complete existing list of
parameters to be passed to \code{func} (\code{TRUE}), or to simply add
the new parameters to the existing list and only replace overlapping
parameters (\code{FALSE}). (default = \code{FALSE})}
}
\value{
Modified \code{\link[=BenchDesign-class]{BenchDesign}} object with single method
parameters modified.
}
\description{
Takes a \code{\link[=BenchDesign-class]{BenchDesign}} object, the name of an
existing method, and new parameter specifications,
and returns a modified \code{\link[=BenchDesign-class]{BenchDesign}} object with
the specified changes.
}
\examples{
## empty BenchDesign
bench <- BenchDesign()

## add method
bench <- addMethod(bench, label = "qv",
                   func = qvalue::qvalue,
                   post = function(x) { x$qvalue },
                   meta = list(note = "storey's q-value"),
                   params = rlang::quos(p = pval))

## modify method 'meta' property of 'qv' method
bench <- modifyMethod(bench, label = "qv",
                      params = rlang::quos(bd.meta =
                           list(note = "Storey's q-value")))

## verify that method has been updated
printMethod(bench, "qv")

}
\seealso{
\code{\link{addMethod}}, \code{\link{expandMethod}}, \code{\link{dropMethod}}
}
\author{
Patrick Kimes
}
