% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifyMethod.R
\name{.modmethod}
\alias{.modmethod}
\title{Modify BenchDesign Method}
\usage{
.modmethod(m, q, .overwrite)
}
\arguments{
\item{m}{method}

\item{q}{quosure list of new parameters}

\item{.overwrite}{logical whether to overwrite parameters}
}
\value{
modified method.
}
\description{
Given a method defined in a BenchDesign, this helper function
returns a modified method with new parameters defined as a
list of quosures.
}
\author{
Patrick Kimes
}
\keyword{internal}
