% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/SummarizedBenckmark-accessors.R
\docType{methods}
\name{performanceMetrics<-}
\alias{performanceMetrics<-}
\alias{performanceMetrics<-,SummarizedBenchmark,SimpleList-method}
\title{Set performance metrics in SummarizedBenchmark object}
\usage{
performanceMetrics(object, ...) <- value


  \S4method{performanceMetrics}{SummarizedBenchmark,SimpleList}(object) <- value
}
\arguments{
\item{object}{a \code{\link[=SummarizedBenchmark-class]{SummarizedBenchmark}} object.}

\item{...}{futher arguments, perhaps used by methods.}

\item{value}{a SimpleList of the same length as the number of assays.}
}
\value{
Silently, the newly specified SimpleList of performance metric lists.
}
\description{
Replaces the list of performance metrics in a
\code{\link[=SummarizedBenchmark-class]{SummarizedBenchmark}} object with
a new list of performance metric lists.
}
\examples{
data(sb)
performanceMetrics(sb)
performanceMetrics(sb) <- SimpleList(qvalue = list(), logFC = list())

}
\seealso{
\code{\link{addPerformanceMetric}}, \code{\link{estimatePerformanceMetrics}}, \code{\link{performanceMetrics}}
}
\author{
Alejandro Reyes
}
