% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/SummarizedBenckmark-accessors.R
\docType{methods}
\name{performanceMetrics}
\alias{performanceMetrics}
\alias{performanceMetrics,SummarizedBenchmark-method}
\title{Get performance metrics in SummarizedBenchmark object}
\usage{
performanceMetrics(object, ...)

\S4method{performanceMetrics}{SummarizedBenchmark}(object, assay = NULL)
}
\arguments{
\item{object}{a \code{\link[=SummarizedBenchmark-class]{SummarizedBenchmark}} object.}

\item{...}{futher arguments, perhaps used by methods.}

\item{assay}{a character string indicating an assay name.}
}
\value{
A SimpleList with one element for each assay. Each element of the list contains a list of performance metric functions.
}
\description{
Given a \code{\link[=SummarizedBenchmark-class]{SummarizedBenchmark}} object,
returns a list of lists of performance metrics that have been defined for
each assay. Optionally, if \code{assay =} is specified, performance metrics
for only the specified subset of specified assays are returned.
}
\examples{
data(sb)
performanceMetrics(sb)
performanceMetrics(sb, assay = "qvalue")

}
\seealso{
\code{\link{addPerformanceMetric}}, \code{\link{estimatePerformanceMetrics}}
}
\author{
Alejandro Reyes
}
