% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{plotROC}
\alias{plotROC}
\title{Plotting ROC curves}
\usage{
plotROC(object, assay = "qvalue")
}
\arguments{
\item{object}{A \code{\link{SummarizedBenchmark}} object.}

\item{assay}{An assay name.}
}
\value{
An \code{\link{SummarizedBenchmark}} object with performance functions.
}
\description{
This function inputs a \code{\link{SummarizedBenchmark}} object, looks
for an assay called 'qvalue' and plots receiver operating characteristic curves
for each of the methods to benchmark.
}
\examples{
\dontrun{
data( sb )
plotROC( sb )
}

}
\author{
Alejandro Reyes
}
