% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchDesign-comparisons.R
\name{printCompareBenchDesigns}
\alias{printCompareBenchDesigns}
\title{Print Comparison of BenchDesign objects}
\usage{
printCompareBenchDesigns(x, y)
}
\arguments{
\item{x}{a \code{BenchDesign} object}

\item{y}{a \code{BenchDesign} object}
}
\value{
logical value indicating whether the two objects have
methods producing the same meta data and data with the
same MD5 hashes.
}
\description{
Simple comparison of two BenchDesign objects based on
methods meta data and data MD5 hashs.
}
\author{
Patrick Kimes
}
