% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchDesign-comparisons.R
\name{printUpdateBench}
\alias{printUpdateBench}
\title{Print Comparison of BenchDesign objects}
\usage{
printUpdateBench(sb, bd, version = FALSE)
}
\arguments{
\item{x}{a BenchDesign or SummarizedBenchmark object}

\item{y}{a BenchDesign or SummarizedBenchmark object}
}
\value{
logical value indicating whether the two objects have
methods producing the same meta data and data with the
same MD5 hashes.
}
\description{
Simple comparison of two BenchDesign objects based on
methods meta data and data MD5 hashs.
}
\author{
Patrick Kimes
}
