% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showBMethod.R
\name{showBMethod}
\alias{showBMethod}
\alias{showBMethods}
\title{Pretty print methods in a BenchDesign}
\usage{
showBMethod(b, n)

showBMethods(b)
}
\arguments{
\item{b}{BenchDesign object.}

\item{n}{name of a method in the BenchDesign to show.}
}
\value{
Brief description is returned to console.
}
\description{
Print out details about a method included in
the BenchDesign. The \code{showBMethods} function is just a wrapper
to call \code{showBMethod} on all methods in the BenchDesign.
}
\examples{
\dontrun{
bd <- BenchDesign()
showBMethods(bd)
}

}
\author{
Patrick Kimes

Patrick Kimes
}
