% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/SummarizedBenckmark-accessors.R
\docType{methods}
\name{SummarizedBenchmark-accessors}
\alias{SummarizedBenchmark-accessors}
\alias{groundTruths}
\alias{groundTruths<-}
\alias{assayNames<-,SummarizedBenchmark,character-method}
\alias{assayNames}
\alias{assayNames,SummarizedBenchmark-method}
\alias{mcols<-,SummarizedBenchmark-method}
\alias{groundTruths,SummarizedBenchmark-method}
\alias{groundTruths}
\alias{groundTruths<-,SummarizedBenchmark-method}
\alias{groundTruths<-,SummarizedBenchmark-method}
\alias{BDMethodList,SummarizedBenchmark-method}
\alias{BenchDesign,SummarizedBenchmark-method}
\title{Accessor and replacement functions for the slots of a SummarizedBenchmark object.}
\usage{
groundTruths(object, ...)

groundTruths(object, ...) <- value

\S4method{assayNames}{SummarizedBenchmark,character}(x, ...) <- value

\S4method{mcols}{SummarizedBenchmark}(x, ...) <- value

\S4method{groundTruths}{SummarizedBenchmark}(object, ...)

\S4method{groundTruths}{SummarizedBenchmark}(object, ...) <- value

\S4method{BDMethodList}{SummarizedBenchmark}(..., x = NULL)

\S4method{BenchDesign}{SummarizedBenchmark}(methods, data)
}
\arguments{
\item{object}{a \code{SummarizedBenchmark} object.}

\item{...}{Futher arguments, perhaps used by methods}

\item{value}{A character vector.}

\item{x}{A \code{SummarizedBenchmark} object.}

\item{methods}{A \code{SummarizedBenchmark} object.}

\item{data}{A data set to be used.}
}
\value{
Either a \code{SummarizedBenchmark} object or a slot of the \code{SummarizedBenchmark} object.
}
\description{
Accessor and replacement functions for the slots of a SummarizedBenchmark object.
}
\examples{

data( sb )
assayNames( sb )[2] <- "log2FC"

}
\seealso{
\code{\link{performanceMetrics}}
}
\author{
Alejandro Reyes
}
