% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SummarizedBenchmark-class}
\alias{SummarizedBenchmark-class}
\title{SummarizedBenchmark class documentation}
\description{
Extension of the \code{\link{RangedSummarizedExperiment}} to
store the output of different methods intended for the same purpose
in a given dataset. For example, a differential expression analysis could be
done using \pkg{limma}-voom, \pkg{edgeR} and \pkg{DESeq2}. The
SummarizedBenchmark class provides a framework that is useful to store, benckmark and
compare results.
}
\section{Slots}{

\describe{
\item{\code{performanceMetrics}}{A \code{\link{SimpleList}} of the same length
as the number of \code{\link{assays}} containing performance
functions to be compared with the ground truths.}
}}

\examples{

## loading the example data from iCOBRA
library(iCOBRA)
data(cobradata_example)

## a bit of data wrangling and reformatting
assays <- list(
    qvalue=cobradata_example@padj,
    logFC=cobradata_example@score )
assays[["qvalue"]]$DESeq2 <- p.adjust(cobradata_example@pval$DESeq2, method="BH")
groundTruth <- DataFrame( cobradata_example@truth[,c("status", "logFC")] )
colnames(groundTruth) <- names( assays )
colData <- DataFrame( method=colnames(assays[[1]]) )
groundTruth <- groundTruth[rownames(assays[[1]]),]

## constructing a SummarizedBenchmark object
sb <- SummarizedBenchmark(
    assays=assays, colData=colData,
    groundTruth=groundTruth )

}
\author{
Alejandro Reyes
}
