% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDMethod-tidy.R
\name{tidyBDMethod}
\alias{tidyBDMethod}
\title{Tidy BDMethod Data}
\usage{
tidyBDMethod(bdm, dat = NULL)
}
\arguments{
\item{bdm}{a BDMethod object}

\item{dat}{optional data object to use when evaluating any
unevaluated expressionsin \code{bdm} meta data.
(default = NULL)}
}
\value{
A named vector of meta data for the specified BDMethod.
}
\description{
A helper function to create tabular info for a single
BDMethod object.
}
\author{
Patrick Kimes
}
