% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformanceMetrics.R
\name{tidyUpMetrics}
\alias{tidyUpMetrics}
\title{Tidy up performance metrics in SummarizedBenchmark object}
\usage{
tidyUpMetrics(object)
}
\arguments{
\item{object}{A \code{\link{SummarizedBenchmark}} object.}
}
\value{
A tidy \code{data.frame}
}
\description{
This function takes as input a \code{SummarizedBenchmark} object, extracts the
estimated performance metrics and reformats them into a long-formated data frame.
}
\examples{
data( "sb", package="SummarizedBenchmark" )
sb <- estimateMetricsForAssay( sb, assay="qvalue", evalMetric="rejections",
    evalFunction=function( query, truth, alpha=0.1 ){
        sum( query < alpha )
    },
    addColData=TRUE )
tidyUpMetrics( sb )

}
\seealso{
\code{\link{estimateMetrics}}
}
\author{
Alejandro Reyes
}
