% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-soneson2016.R
\docType{data}
\name{truthdat}
\alias{truthdat}
\alias{txi}
\title{Data objects containing counts and ground truths of the simulations from Soneson et al., 2016.}
\description{
Data objects containing counts and ground truths of the simulations from Soneson et al., 2016.
}
\keyword{data}
