% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedBenckmark-class.R
\docType{methods}
\name{Accessors}
\alias{Accessors}
\alias{assayNames}
\alias{assayNames,SummarizedBenchmark-method}
\alias{assayNames<-,SummarizedBenchmark,character-method}
\alias{Accessors}
\alias{mcols<-,SummarizedBenchmark-method}
\title{Accessor and replacement of the assay names of a SummarizedBenchmark object.}
\usage{
\S4method{assayNames}{SummarizedBenchmark,character}(x, ...) <- value

\S4method{mcols}{SummarizedBenchmark}(x, ...) <- value
}
\arguments{
\item{x}{a \code{SummarizedBenchmark} object.}

\item{...}{Futher arguments, perhaps used by methods}

\item{value}{A character vector}
}
\description{
Accessor and replacement of the assay names of a SummarizedBenchmark object.
}
\examples{

data( sb )
assayNames( sb )[2] <- "log2FC"

}
\seealso{
\code{\link{performanceMetrics}}
}
\author{
Alejandro Reyes
}
