% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefaultMetrics.R
\name{addDefaultMetrics}
\alias{addDefaultMetrics}
\title{Add default performance metrics to a \code{\link{SummarizedBenchmark}} object.}
\usage{
addDefaultMetrics(object)
}
\arguments{
\item{object}{A \code{\link{SummarizedBenchmark}} object.}
}
\value{
A \code{\link{SummarizedBenchmark}} object with performance functions.
}
\description{
This function adds predefined performance metrics to a \code{\link{SummarizedBenchmark}}
object, according to the assay name. For example, if an assay is named "qvalue",
it will add a the performance metric "rejections".
}
\details{
This function adds a set of predefined performance metrics according
to the names of the assays. Specifically, if an assay is called "qvalue" it will
add the "rejections" metric by default. Furthermore, if the qvalue assay has
information about the ground truths in the rowData, the metrics TPR (true
positive rate), "TNR" (true negative rate), "FPR" (false positive rate) and
"FNR" (false negative rate) will be added.
}
\examples{

data( sb )
sb <- addDefaultMetrics( sb )

}
\author{
Alejandro Reyes
}
