% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMethod.R
\name{addMethod}
\alias{addMethod}
\title{Add new method to BenchDesign object}
\usage{
addMethod(b, blabel, bfunc, bpost = NULL, ...)
}
\arguments{
\item{b}{BenchDesign object}

\item{blabel}{character name for the method}

\item{bfunc}{primary function to be benchmarked}

\item{bpost}{optional post-processing function that takes output of
\code{bfunc} e.g. a getter method, ignored if NULL
(default = NULL)}

\item{...}{named \code{parameter = value} pairs to be passed to
\code{func}}
}
\description{
Add new method to BenchDesign object
}
\author{
Patrick Kimes
}
