% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-soneson2016.R
\docType{data}
\name{allSB}
\alias{allSB}
\title{SummarizedBenchmark object of isoform quantification results}
\description{
SummarizedBenchmark object of isoform quantification results
}
\keyword{data}
