% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedBenckmark-accessors.R
\docType{methods}
\name{assayNames-setter}
\alias{assayNames-setter}
\title{Set assay names in SummarizedBenchmark object}
\usage{
\S4method{assayNames}{SummarizedBenchmark,character}(x, ...) <- value
}
\arguments{
\item{x}{A \code{\link[=SummarizedBenchmark-class]{SummarizedBenchmark}} object.}

\item{...}{Futher arguments, perhaps used by methods.}

\item{value}{A character vector.}
}
\value{
Modfied \code{\link[=SummarizedBenchmark-class]{SummarizedBenchmark}} object.
}
\description{
Modifies the assay names of a \code{\link[=SummarizedBenchmark-class]{SummarizedBenchmark}}
object.
}
\examples{
data(sb)
assayNames(sb)[2] <- "log2FC"

}
\seealso{
\code{\link[=SummarizedBenchmark-class]{SummarizedBenchmark}}
}
\author{
Alejandro Reyes
}
