% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefaultMetrics.R
\name{availableMetrics}
\alias{availableMetrics}
\title{availableMetrics}
\usage{
availableMetrics()
}
\value{
A data.frame summarizing the default performance metrics provided in this package.
}
\description{
This function returns a data frame summarizing the default performance metrics provided in this package.
The data.frame contains three columns, `functions` is the name of the performance metric, `description`
is longer description of the performance metric and `requiredTruth` is logical depending on whether the
performance metrics require ground truths.
}
\examples{
availableMetrics()

}
