% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/BDData-show.R
\docType{class}
\name{BDData-class}
\alias{BDData-class}
\alias{show,BDData-method}
\title{BDMethod class}
\usage{
\S4method{show}{BDData}(object)
}
\arguments{
\item{object}{BDData object to show}
}
\description{
Container for individual methods to be compared as part of
a BenchDesign object.
}
\section{Slots}{

\describe{
\item{\code{data}}{a list or MD5 hash of the data.}

\item{\code{type}}{a character string indicating whether the data slot
contains the 'data' or a 'md5hash' of the data.}
}}

\author{
Patrick Kimes
}
