% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/BDData-constructor.R
\docType{class}
\name{BDData-class}
\alias{BDData-class}
\alias{BDData}
\alias{BDData,ANY-method}
\title{BDMethod class}
\usage{
BDData(data)

\S4method{BDData}{ANY}(data)
}
\arguments{
\item{data}{a list object of data or MD5 hash string}
}
\value{
BDData object
}
\description{
Container for individual methods to be compared as part of
a BenchDesign object.

Initialized a new BDData object for benchmarking methods.
}
\section{Slots}{

\describe{
\item{\code{data}}{a list or MD5 hash of the data.}

\item{\code{type}}{a character string indicating whether the data slot
contains the 'data' or a 'md5hash' of the data.}
}}

\author{
Patrick Kimes

Patrick Kimes
}
