% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BenchDesign-accessors.R
\docType{methods}
\name{BDData-setter}
\alias{BDData-setter}
\alias{BDData<-,BenchDesign,BDDataOrNULL-method}
\title{BDData slot setter}
\usage{
BDData(x) <- value

\S4method{BDData}{BenchDesign,BDDataOrNULL}(x) <- value
}
\arguments{
\item{x}{BenchDesign object.}
}
\value{
modified BenchDesign object
}
\description{
Method to set BDData slot in objects.
}
\author{
Patrick Kimes
}
