% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BDData-constructor.R
\docType{methods}
\name{BDData}
\alias{BDData}
\alias{BDData,ANY-method}
\alias{BDData,BenchDesign-method}
\alias{BDData,SummarizedBenchmark-method}
\alias{BDData,BDData-method}
\title{Create a new BDData object}
\usage{
BDData(data)

\S4method{BDData}{ANY}(data)

\S4method{BDData}{BenchDesign}(data)

\S4method{BDData}{SummarizedBenchmark}(data)

\S4method{BDData}{BDData}(data)
}
\arguments{
\item{data}{a list object of data or MD5 hash string}
}
\value{
BDData object
}
\description{
Initialized a new BDData object for benchmarking methods.
}
\author{
Patrick Kimes
}
