% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BDMethod-class}
\alias{BDMethod-class}
\title{BDMethod class}
\description{
Container for individual methods to be compared as part of a benchmark
experiment defined in a BenchDesign object. In the SummarizedBenchmark
framework, methods are defined by a unique combination of functions,
parameters, and any relevant meta data.
 
New BDMethod objects can be created using the \code{\link{BDMethod}}
constructor.
}
\section{Slots}{

\describe{
\item{\code{f}}{a function to be benchmarked}

\item{\code{fc}}{a captured expression of the function \code{f}}

\item{\code{params}}{a list of quosures specifying function parameters}

\item{\code{post}}{a list of functions to be applied to the output of \code{f}}

\item{\code{meta}}{a list of meta data}
}}

\seealso{
\code{\link{BDMethod}}, \code{\link{BenchDesign-class}}, \code{\link{BDMethodList-class}}, \code{\link{BDData-class}}
}
\author{
Patrick Kimes
}
