% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/BDMethod-constructor.R,
%   R/BenchDesign-show.R
\docType{class}
\name{BDMethod-class}
\alias{BDMethod-class}
\alias{BDMethod}
\alias{BDMethod,quosure-method}
\alias{show,BDMethod-method}
\title{BDMethod class}
\usage{
BDMethod(f, params = rlang::quos(), post = NULL, meta = NULL)

\S4method{BDMethod}{quosure}(f, params = rlang::quos(), post = NULL,
  meta = NULL)

\S4method{show}{BDMethod}(object)
}
\arguments{
\item{f}{a function to be benchmarked}

\item{params}{a list of quosures specifying function parameters}

\item{post}{a list of functions to be applied to the output of \code{f}}

\item{meta}{a list of meta data}
}
\value{
BDMethod object
}
\description{
Container for individual methods to be compared as part of
a BenchDesign object.

Initializes a new BenchDesign object for benchmarking methods.
}
\section{Slots}{

\describe{
\item{\code{f}}{a function to be benchmarked}

\item{\code{fc}}{a captured expression of the function \code{f}}

\item{\code{params}}{a list of quosures specifying function parameters}

\item{\code{post}}{a list of functions to be applied to the output of \code{f}}

\item{\code{meta}}{a list of meta data}
}}

\author{
Patrick Kimes

Patrick Kimes
}
