% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BDMethodList-accessors.R,
%   R/BenchDesign-accessors.R
\docType{methods}
\name{BDMethod-setter}
\alias{BDMethod-setter}
\alias{BDMethod<-,BDMethodList,character,BDMethod-method}
\alias{BDMethod<-,BDMethodList,character,NULL-method}
\alias{BDMethod<-,BenchDesign,character,BDMethod-method}
\alias{BDMethod<-,BenchDesign,character,NULL-method}
\title{BDMethod slot setter}
\usage{
BDMethod(x, i) <- value

\S4method{BDMethod}{BDMethodList,character,BDMethod}(x, i) <- value

\S4method{BDMethod}{BDMethodList,character,`NULL`}(x, i) <- value

\S4method{BDMethod}{BenchDesign,character,BDMethod}(x, i) <- value

\S4method{BDMethod}{BenchDesign,character,`NULL`}(x, i) <- value
}
\arguments{
\item{x}{BenchDesign object.}
}
\value{
modified BenchDesign object
}
\description{
Method to set BDMethod slot in objects.
}
\author{
Patrick Kimes
}
