% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BDMethod-constructor.R
\docType{methods}
\name{BDMethod}
\alias{BDMethod}
\alias{BDMethod}
\alias{BDMethod,quosure-method}
\alias{BDMethod,function-method}
\title{Create a new BDMethod}
\usage{
BDMethod(x, params = rlang::quos(), post = NULL, meta = NULL, ...)

\S4method{BDMethod}{quosure}(x, params = rlang::quos(), post = NULL,
  meta = NULL, ...)

\S4method{BDMethod}{`function`}(x, params = rlang::quos(), post = NULL,
  meta = NULL, ...)
}
\arguments{
\item{x}{main method function or function quosure}

\item{params}{list of quosures specifying function parameters.
(default = code{rlang::quos()})}

\item{post}{list of functions to be applied to the output of \code{x}.
(default = NULL)}

\item{meta}{list of meta data. (default = NULL)}

\item{...}{other parameters.}
}
\value{
BDMethod object
}
\description{
Initializes a new BenchDesign method (BDMethod) object for benchmarking methods.
}
