% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BDMethod-constructor.R,
%   R/BDMethodList-accessors.R, R/BenchDesign-accessors.R
\docType{methods}
\name{BDMethod}
\alias{BDMethod}
\alias{BDMethod,quosure-method}
\alias{BDMethod,function-method}
\alias{BDMethod,BDMethodList-method}
\alias{BDMethod,BenchDesign-method}
\title{Create a new BDMethod object}
\usage{
BDMethod(x, params = rlang::quos(), post = NULL, meta = NULL, ...)

\S4method{BDMethod}{quosure}(x, params = rlang::quos(), post = NULL,
  meta = NULL, ...)

\S4method{BDMethod}{`function`}(x, params = rlang::quos(), post = NULL,
  meta = NULL, ...)

\S4method{BDMethod}{BDMethodList}(x, i = 1)

\S4method{BDMethod}{BenchDesign}(x, i = 1)
}
\arguments{
\item{x}{main method function or function quosure. Alternative, may be a
\code{\link[=BDMethodList-class]{BDMethodList}} or
\code{\link[=BenchDesign-class]{BenchDesign}} object from which
the \code{\link[=BDMethod-class]{BDMethod}} should be
extracted.}

\item{params}{list of quosures specifying function parameters.
(default = \code{rlang::quos()})}

\item{post}{list of functions to be applied to the output of \code{x}.
(default = NULL)}

\item{meta}{list of metadata. (default = NULL)}

\item{...}{other parameters.}

\item{i}{integer index or character name of \code{\link[=BDMethod-class]{BDMethod}} in
\code{\link[=BDMethodList-class]{BDMethodList}} or
\code{\link[=BenchDesign-class]{BenchDesign}} object.}
}
\value{
BDMethod object
}
\description{
Initializes a new BenchDesign method object for benchmarking.

New \code{\link[=BDMethod-class]{BDMethod}} objects are typically not directly constructed as they
have limited use outside of \code{\link[=BenchDesign-class]{BenchDesign}} objects. Instead,
methods in a \code{\link[=BenchDesign-class]{BenchDesign}} object are more commonly created, modified or
removed using
function calls on the \code{\link[=BenchDesign-class]{BenchDesign}}, e.g. using \code{\link{addMethod}} to
add a new method object.

The constructor can also be used to access \code{\link[=BDMethod-class]{BDMethod}} objects stored in
\code{\link[=BDMethodList-class]{BDMethodList}} and \code{\link[=BenchDesign-class]{BenchDesign}} objects.
}
\examples{
## create a simple BDMethod
bdm1 <- BDMethod(x = base::mean)

## create a more complex BDMethod
bdm2 <- BDMethod(x = function(x) { x^2 }, post = base::sqrt,
                 meta = list(note = "simple example"))

## construct a BenchDesign with the BDMethod objects
bd <- BenchDesign(method1 = bdm1, method2 = bdm2)

## access a BDMethod in the BenchDesign
BDMethod(bd, "method1")

}
\seealso{
\code{\link{BDMethod-class}}, \code{\link{BenchDesign}}, \code{\link{BDMethodList}}
}
\author{
Patrick Kimes
}
