% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/BDMethodList-show.R
\docType{class}
\name{BDMethodList-class}
\alias{BDMethodList-class}
\alias{show,BDMethodList-method}
\title{BDMethodList class}
\usage{
\S4method{show}{BDMethodList}(object)
}
\arguments{
\item{object}{BDMethodList object to show}
}
\description{
Extension of the SimpleList class to contain a list of BDMethod
objects.
}
\author{
Patrick Kimes
}
