% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BDMethodList-class}
\alias{BDMethodList-class}
\title{BDMethodList class}
\description{
Extension of the SimpleList class to contain a list of BDMethod
objects. The class serves as the primary container for the set of
methods in the BenchDesign class.

New BDMethodList objects can be created using the
\code{\link{BDMethodList}} constructor.
}
\seealso{
\code{\link{BDMethodList}}, \code{\link{BenchDesign-class}}, \code{\link{BDMethod-class}}, \code{\link{BDData-class}}
}
\author{
Patrick Kimes
}
