% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDMethod-constructor.R
\docType{methods}
\name{BDMethodList}
\alias{BDMethodList}
\alias{BDMethodList,ANY-method}
\title{Create a new BDMethodList}
\usage{
BDMethodList(..., object = NULL)

\S4method{BDMethodList}{ANY}(..., object = NULL)
}
\arguments{
\item{...}{a named list of BDMethod objects}
}
\value{
BDMethodList object
}
\description{
Initialized a new SimpleList of BenchDesign method (BDMethod) objects.
}
\examples{
BDMethodList()

}
\author{
Patrick Kimes
}
