% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BDMethodList-class}
\alias{BDMethodList-class}
\title{BDMethodList class}
\description{
Extension of the SimpleList class to contain a list of BDMethod
objects.
}
\author{
Patrick Kimes
}
