% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BenchDesign-accessors.R
\docType{methods}
\name{BDMethodList-setter}
\alias{BDMethodList-setter}
\alias{BDMethodList<-,BenchDesign,BDMethodList-method}
\title{BDMethodList slot setter}
\usage{
BDMethodList(x) <- value

\S4method{BDMethodList}{BenchDesign,BDMethodList}(x) <- value
}
\arguments{
\item{x}{BenchDesign object.}
}
\value{
modified BenchDesign object
}
\description{
Method to set BDMethodList slot in objects.
}
\author{
Patrick Kimes
}
