% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BDMethodList-constructor.R,
%   R/BenchDesign-accessors.R, R/SummarizedBenckmark-accessors.R
\docType{methods}
\name{BDMethodList}
\alias{BDMethodList}
\alias{BDMethodList,ANY-method}
\alias{coerce}
\alias{BDMethodList,BenchDesign-method}
\alias{BDMethodList,SummarizedBenchmark-method}
\title{Create a new BDMethodList object}
\usage{
BDMethodList(..., x = NULL)

\S4method{BDMethodList}{ANY}(..., x = NULL)

\S4method{BDMethodList}{BenchDesign}(..., x = NULL)

\S4method{BDMethodList}{SummarizedBenchmark}(..., x = NULL)
}
\arguments{
\item{...}{a named list of \code{\link[=BDMethod-class]{BDMethod}} objects}

\item{x}{a \code{\link[=BenchDesign-class]{BenchDesign}} or
\code{\link[=SummarizedBenchmark-class]{SummarizedBenchmark}} object to extract
the list of methods from. (default = NULL)}
}
\value{
BDMethodList object
}
\description{
Initializes a new SimpleList of BenchDesign method (\code{\link[=BDMethod-class]{BDMethod}}) objects.

Similar to \code{\link[=BDMethod-class]{BDMethod}} objects, \code{\link[=BDMethodList-class]{BDMethodList}}
typically do not need to be
directly constructed. Because the list of methods is only useful as art of a
\code{\link[=BenchDesign-class]{BenchDesign}} object, it is more common to simply manipulate the
list of methods through calls to the corresponding \code{\link[=BenchDesign-class]{BenchDesign}}, e.g.
\code{\link{addMethod}} to add a new method to the list.

The constructor can also be used to access the \code{\link[=BDMethodList-class]{BDMethodList}} list
of methods in a \code{\link[=BenchDesign-class]{BenchDesign}} object.
}
\examples{
## construct an empty list
bdml <- BDMethodList()

## construct a list with BDMethod objects
bdml <- BDMethodList(m_method = BDMethod(base::mean),
                     s_method = BDMethod(function(x) { x^2 }))
bdml

## construct a BenchDesign with a BDMethodList
bd <- BenchDesign(methods = bdml)

## access the BDMethodList in the BenchDesign
BDMethodList(bd)

}
\seealso{
\code{\link{BDMethodList-class}}, \code{\link{BenchDesign}}, \code{\link{BDMethod}}
}
\author{
Patrick Kimes
}
