% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/BenchDesign-show.R
\docType{class}
\name{BenchDesign-class}
\alias{BenchDesign-class}
\alias{show,BenchDesign-method}
\title{BenchDesign class}
\usage{
\S4method{show}{BenchDesign}(object)
}
\arguments{
\item{object}{BenchDesign object to show}
}
\description{
Container for benchmark methods and data.
}
\section{Slots}{

\describe{
\item{\code{data}}{a list containing the data to be used in the benchmark.}

\item{\code{methods}}{a list of \code{BDMethods} to be compared in the benchmark.}
}}

\author{
Patrick Kimes
}
