% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/BenchDesign-constructor.R,
%   R/BenchDesign-show.R
\docType{class}
\name{BenchDesign-class}
\alias{BenchDesign-class}
\alias{BenchDesign}
\alias{BenchDesign,ANY-method}
\alias{show,BenchDesign-method}
\title{BenchDesign class}
\usage{
BenchDesign(..., methods = NULL, data = NULL)

\S4method{BenchDesign}{ANY}(..., methods = NULL, data = NULL)

\S4method{show}{BenchDesign}(object)
}
\arguments{
\item{...}{named set of \code{BDMethod} objects.}

\item{methods}{named set of \code{BDMethod} objects as a list. (default = NULL)}

\item{data}{optional data.frame or other list object to be
used in the benchmark. (default = NULL)}
}
\value{
\code{BenchDesign} object.
}
\description{
Container for benchmark methods and data.

Initializes a new BenchDesign object for benchmarking methods.
}
\section{Slots}{

\describe{
\item{\code{data}}{a list containing the data to be used in the benchmark.}

\item{\code{methods}}{a list of \code{BDMethods} to be compared in the benchmark.}
}}

\examples{
## with no input
bd <- BenchDesign()

## with toy data.frame
df <- data.frame(x1 = rnorm(20), y1 = rnorm(20))
bd <- BenchDesign(data = df)

}
\author{
Patrick Kimes

Patrick Kimes
}
