% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchDesign-show.R
\docType{methods}
\name{show,BenchDesign-method}
\alias{show,BenchDesign-method}
\title{Show BenchDesign object}
\usage{
\S4method{show}{BenchDesign}(object)
}
\arguments{
\item{object}{BenchDesign object to show}
}
\value{
Print description of BenchDesign object to console
}
\description{
Show BenchDesign object
}
