% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchDesign.R
\name{BenchDesign}
\alias{BenchDesign}
\title{create a new BenchDesign}
\usage{
BenchDesign(bdata = NULL)
}
\arguments{
\item{bdata}{data to be used for benchmarking methods
(default = NULL)}
}
\value{
a BenchDesign object
}
\description{
Initializes a new BenchDesign object.
}
\author{
Patrick Kimes
}
