% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BenchDesign-constructor.R,
%   R/SummarizedBenckmark-accessors.R
\docType{methods}
\name{BenchDesign}
\alias{BenchDesign}
\alias{BenchDesign,ANY-method}
\alias{BenchDesign,SummarizedBenchmark-method}
\title{Create a new BenchDesign object}
\usage{
BenchDesign(..., methods = NULL, data = NULL)

\S4method{BenchDesign}{ANY}(..., methods = NULL, data = NULL)

\S4method{BenchDesign}{SummarizedBenchmark}(methods, data)
}
\arguments{
\item{...}{named set of \code{BDMethod} objects and/or unnamed
\code{BenchDesign} objects. Only the methods of any \code{BenchDesign}
object will be used, and the data slot of the objects will be
ignored.}

\item{methods}{named set of \code{BDMethod} objects and/or unnamed
\code{BenchDesign} objects as a list. (default = NULL)}

\item{data}{optional data.frame or other list object to be
used in the benchmark. (default = NULL)}
}
\value{
\code{BenchDesign} object.
}
\description{
Initializes a new BenchDesign object of benchmarking methods and data.

The BenchDesign class serves as the core container
for methods and data used for benchmarking in the SummarizedBenchmark package. The object
can be initialized with a list of methods ot be benchmarked, a default benchmarking data set,
both or neither. Methods must be passed to the constructor as \code{BDMethod} or \code{BDMethodList}
objects.

The constructor can also be used to access the BenchDesign stored in a SummarizedBenchmark
object.
}
\examples{
## with no input
bd <- BenchDesign()

## with data - data must be a named argument
datadf <- data.frame(pval = runif(20), x1 = rnorm(20))
bd <- BenchDesign(data = datadf)

## with two methods and data
method_bh <- BDMethod(stats::p.adjust, params = rlang::quos(p = pval, method = "BH"))
method_bf <- BDMethod(stats::p.adjust, params = rlang::quos(p = pval, method = "bonferroni"))
bd <- BenchDesign(bh = method_bh, bonf = method_bf,
                  data = datadf)

## with BDMethodList and data
bdml <- BDMethodList(bh = methods_bh, bonf = method_bf)
bd <- BenchDesign(methods = bdml, data = datadf)

}
\seealso{
\code{\link{BenchDesign-class}}, \code{\link{BDMethod}}, \code{\link{BDMethodList}}
}
\author{
Patrick Kimes
}
