% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchDesign.R
\name{BenchDesign}
\alias{BenchDesign}
\title{Create a new BenchDesign}
\usage{
BenchDesign(bdata = NULL)
}
\arguments{
\item{bdata}{optional data.frame or other list object to be
used in the benchmark. (default = NULL)}
}
\value{
BenchDesign object
}
\description{
Initializes a new BenchDesign object for benchmarking methods.
}
\examples{
## with no input
bd <- BenchDesign()

## with toy data.frame
df <- data.frame(x1 = rnorm(20), y1 = rnorm(20))
bd <- BenchDesign(df)

}
\author{
Patrick Kimes
}
