% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BenchDesign-constructor.R,
%   R/SummarizedBenckmark-accessors.R
\docType{methods}
\name{BenchDesign}
\alias{BenchDesign}
\alias{BenchDesign,ANY-method}
\alias{BenchDesign,SummarizedBenchmark-method}
\title{Create a new BenchDesign object}
\usage{
BenchDesign(..., methods = NULL, data = NULL)

\S4method{BenchDesign}{ANY}(..., methods = NULL, data = NULL)

\S4method{BenchDesign}{SummarizedBenchmark}(methods, data)
}
\arguments{
\item{...}{named set of \code{BDMethod} objects and/or unnamed
\code{BenchDesign} objects. Only the methods of any \code{BenchDesign}
object will be used, and the data slot of the objects will be
ignored.}

\item{methods}{named set of \code{BDMethod} objects and/or unnamed
\code{BenchDesign} objects as a list. (default = NULL)}

\item{data}{optional data.frame or other list object to be
used in the benchmark. (default = NULL)}
}
\value{
\code{BenchDesign} object.
}
\description{
Initializes a new BenchDesign object for benchmarking methods.
}
\examples{
## with no input
bd <- BenchDesign()

## with toy data.frame
df <- data.frame(x1 = rnorm(20), y1 = rnorm(20))
bd <- BenchDesign(data = df)

}
\author{
Patrick Kimes
}
