% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildBench.R
\name{buildBench}
\alias{buildBench}
\title{Make benchmark results}
\usage{
buildBench(b)
}
\arguments{
\item{b}{BenchDesign object}
}
\value{
a SummarizedBenchmark with one assay.
}
\description{
Function to evaluate BenchDesign to generate a
SummarizedBenchmark.
}
\author{
Patrick Kimes
}
