% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildBench.R
\name{buildBench}
\alias{buildBench}
\title{Make SummarizedBenchmark from BenchDesign}
\usage{
buildBench(b)
}
\arguments{
\item{b}{BenchDesign object}
}
\value{
SummarizedBenchmark with one assay
}
\description{
Function to evaluate BenchDesign methods on supplied
data set to generate a SummarizedBenchmark.
}
\author{
Patrick Kimes
}
