% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildBench.R
\name{buildBench}
\alias{buildBench}
\title{Make SummarizedBenchmark from BenchDesign}
\usage{
buildBench(b, data = NULL, truthCol = NULL, ptabular = TRUE)
}
\arguments{
\item{b}{BenchDesign object}

\item{data}{data set to be used in place of `bdata` for benchmarking;
ignored if NULL (default = NULL)}

\item{truthCol}{character name of column in data set corresponding to ground
truth values. If specified, column will be added to `groundTruth`
DataFrame of returned SummarizedBenchmark object, and same name
will be used for assay (default = NULL)}

\item{ptabular}{whether to return parameters in tabular form (defualt = TRUE)}
}
\value{
SummarizedBenchmark with one assay
}
\description{
Function to evaluate BenchDesign methods on supplied
data set to generate a SummarizedBenchmark.
}
\author{
Patrick Kimes
}
