% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchDesign-comparisons.R
\name{compareBDData}
\alias{compareBDData}
\title{Compare BDData objects}
\usage{
compareBDData(x, y)
}
\arguments{
\item{x}{a \code{BDData} object}

\item{y}{a \code{BDData} object}
}
\value{
list of two values giving agreement of "data" and "type".
}
\description{
Simple comparison of two BDData objects based on
comparing both type and data hash.
}
