% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchDesign-comparisons.R
\name{compareBDMethod}
\alias{compareBDMethod}
\title{Compare BDMethod objects}
\usage{
compareBDMethod(x, y)
}
\arguments{
\item{x}{a \code{BDMethod} object}

\item{y}{a \code{BDMethod} object}
}
\value{
logical value indicating whether the two objects produced the same
meta data.
}
\description{
Simple comparison of two BDMethod objects based on
meta data.
}
\author{
Patrick Kimes
}
