% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/compareBenchDesigns.R
\docType{methods}
\name{compareBenchDesigns}
\alias{compareBenchDesigns}
\alias{compareBenchDesigns,SummarizedBenchmark,missing-method}
\alias{compareBenchDesigns,SummarizedBenchmark,SummarizedBenchmark-method}
\alias{compareBenchDesigns,SummarizedBenchmark,BenchDesign-method}
\alias{compareBenchDesigns,BenchDesign,SummarizedBenchmark-method}
\alias{compareBenchDesigns,BenchDesign,BenchDesign-method}
\title{Compare BenchDesign objects}
\usage{
compareBenchDesigns(x, y = NULL, ...)

\S4method{compareBenchDesigns}{SummarizedBenchmark,missing}(x, y = NULL,
  ...)

\S4method{compareBenchDesigns}{SummarizedBenchmark,SummarizedBenchmark}(x,
  y = NULL, ...)

\S4method{compareBenchDesigns}{SummarizedBenchmark,BenchDesign}(x,
  y = NULL, ...)

\S4method{compareBenchDesigns}{BenchDesign,SummarizedBenchmark}(x,
  y = NULL, ...)

\S4method{compareBenchDesigns}{BenchDesign,BenchDesign}(x, y = NULL, ...)
}
\arguments{
\item{x}{a SummarizedBenchmark or BenchDesign object}

\item{y}{an optional second SummarizedBenchmark or BenchDesign object
(default = NULL)}

\item{...}{other parameters}
}
\value{
list of comparison results
}
\description{
Comparison of BenchDesign objects and BenchDesign method information
stored in SummarizedBenchmark objects. Inputs can be either BenchDesign
or SummarizedBenchmark objects. If SummarizedBenchmark objects are
specified, the method metadata stored in the \code{colData} will be
used for the comparison. If only a single SummarizedBenchmark object is
specified, the \code{colData} information will be compared with the
BenchDesign object in the \code{BenchDesign} slot of the object.
To compare the \code{BenchDesign} slots of SummarizedBenchmark objects,
the BenchDesigns should be extracted with \code{BenchDesign(sb)} and
passed as inputs (see Examples).
}
\author{
Patrick Kimes
}
