% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropBMethod.R
\name{dropBMethod}
\alias{dropBMethod}
\title{Remove method from BenchDesign object}
\usage{
dropBMethod(b, blabel)
}
\arguments{
\item{b}{BenchDesign object.}

\item{blabel}{Character name of method to be modified.}
}
\value{
Modified BenchDesign object.
}
\description{
This function takes a BenchDesign object and the name of a method
already defined in the object, and returns a reduced BenchDesign
object with the specified method removed.
}
\examples{
## with toy data.frame
df <- data.frame(pval = rnorm(100))
bd <- BenchDesign(df)

## add methods
bd <- addBMethod(bd, blabel = "bonf", bfunc = p.adjust,
                 p = pval, method = "bonferroni")
bd <- addBMethod(bd, blabel = "BH", bfunc = p.adjust,
                 p = pval, method = "BH")

## remove methods
bd <- dropBMethod(bd, blabel = "bonf")

}
\author{
Patrick Kimes
}
