% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropMethod.R
\name{dropMethod}
\alias{dropMethod}
\title{Remove method from BenchDesign object}
\usage{
dropMethod(bd, label)
}
\arguments{
\item{bd}{BenchDesign object.}

\item{label}{Character name of method to be modified.}
}
\value{
Modified BenchDesign object.
}
\description{
This function takes a BenchDesign object and the name of a method
already defined in the object, and returns a reduced BenchDesign
object with the specified method removed.
}
\examples{
## empty BenchDesign
bench <- BenchDesign()

## add methods
bench <- addMethod(bench, label = "bonf", func = p.adjust,
                   params = rlang::quos(p = pval, method = "bonferroni"))
bench <- addMethod(bench, label = "BH", func = p.adjust,
                   params = rlang::quos(p = pval, method = "BH"))

## remove methods
bench <- dropMethod(bench, label = "bonf")

}
\author{
Patrick Kimes
}
