% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/SummarizedBenckmark-accessors.R
\docType{methods}
\name{groundTruths-setter}
\alias{groundTruths-setter}
\alias{groundTruths<-,SummarizedBenchmark-method}
\title{Set ground truths in SummarizedBenchmark object}
\usage{
groundTruths(object, ...) <- value

\S4method{groundTruths}{SummarizedBenchmark}(object, ...) <- value
}
\arguments{
\item{object}{a \code{SummarizedBenchmark} object.}
}
\value{
modified BenchDesign object
}
\description{
Method to set groundTruths in SummarizedBenchmark object.
}
\author{
Patrick Kimes
}
