% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/SummarizedBenckmark-accessors.R
\docType{methods}
\name{groundTruths}
\alias{groundTruths}
\alias{groundTruths,SummarizedBenchmark-method}
\title{groundTruths getter}
\usage{
groundTruths(object, ...)

\S4method{groundTruths}{SummarizedBenchmark}(object, ...)
}
\arguments{
\item{object}{a \code{SummarizedBenchmark} object.}
}
\value{
modified BenchDesign object
}
\description{
Method to get groundTruths in SummarizedBenchmark object.
}
\author{
Patrick Kimes
}
