% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BDData-constructor.R
\docType{methods}
\name{hashBDData}
\alias{hashBDData}
\alias{hashBDData}
\alias{hashBDData,BDData-method}
\alias{hashBDData,BenchDesign-method}
\title{Convert BDData to BDData with MD5 Hash}
\usage{
hashBDData(object)

\S4method{hashBDData}{BDData}(object)

\S4method{hashBDData}{BenchDesign}(object)
}
\arguments{
\item{object}{a \code{BDData} or \code{BenchDesign} object}
}
\value{
an object of the same class as \code{object} with data
converted to a MD5 hash.
}
\description{
Simple converting function to replace data object in
BDData object with MD5 hash value.
}
\author{
Patrick Kimes
}
