% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchDesign.R
\name{is.BenchDesign}
\alias{is.BenchDesign}
\title{Check if object is a BenchDesign}
\usage{
is.BenchDesign(x)
}
\arguments{
\item{x}{object to test}
}
\value{
Logical whether object is a BenchDesign
}
\description{
Check whether an object is a BenchDesign.
}
\author{
Patrick Kimes
}
\keyword{internal}
