% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/SummarizedBenckmark-accessors.R
\docType{methods}
\name{performanceMetrics-setter}
\alias{performanceMetrics-setter}
\alias{performanceMetrics<-,SummarizedBenchmark,SimpleList-method}
\title{performanceMetrics slot setter}
\usage{
performanceMetrics(object, ...) <- value


  \S4method{performanceMetrics}{SummarizedBenchmark,SimpleList}(object) <- value
}
\arguments{
\item{object}{a \code{SummarizedBenchmark} object.}

\item{...}{Futher arguments, perhaps used by methods}

\item{value}{A SimpleList of the same length as the number of assays}

\item{assay}{A character string indicating an assay name}
}
\value{
A SimpleList with one element for each assay. Each element of the list contains a list of performance metric functions.
}
\description{
performanceMetrics slot setter
}
\examples{

data( sb )
performanceMetrics( sb )
performanceMetrics( sb, assay="qvalue" )
performanceMetrics( sb ) <- SimpleList( qvalue=list(), logFC=list() )

}
\seealso{
\code{\link{addPerformanceMetric}}, \code{\link{estimatePerformanceMetrics}}, \code{\link{performanceMetrics}}
}
\author{
Alejandro Reyes
}
