% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/SummarizedBenckmark-accessors.R
\docType{methods}
\name{performanceMetrics}
\alias{performanceMetrics}
\alias{performanceMetrics<-}
\alias{performanceMetrics}
\alias{performanceMetrics,SummarizedBenchmark-method}
\alias{performanceMetrics<-,SummarizedBenchmark,SimpleList-method}
\alias{performanceMetrics,SummarizedBenchmark-method}
\alias{performanceMetrics}
\title{Accessor and replacement functions for the slot 'performanceMetrics' of a SummarizedBenchmark object.}
\usage{
performanceMetrics(object, ...)

performanceMetrics(object, ...) <- value

\S4method{performanceMetrics}{SummarizedBenchmark}(object, assay = NULL)


  \S4method{performanceMetrics}{SummarizedBenchmark,SimpleList}(object) <- value
}
\arguments{
\item{object}{a \code{SummarizedBenchmark} object.}

\item{...}{Futher arguments, perhaps used by methods}

\item{value}{A SimpleList of the same length as the number of assays}

\item{assay}{A character string indicating an assay name}
}
\value{
A SimpleList with one element for each assay. Each element of the list contains a list of performance functions.
}
\description{
Accessor and replacement functions for the slot 'performanceMetrics' of a SummarizedBenchmark object.
}
\examples{

data( sb )
performanceMetrics( sb )
performanceMetrics( sb, assay="qvalue" )
performanceMetrics( sb ) <- SimpleList( qvalue=list(), logFC=list() )

}
\seealso{
\code{\link{addPerformanceMetric}}, \code{\link{estimatePerformanceMetrics}}
}
\author{
Alejandro Reyes
}
