% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{plotMethodsOverlap}
\alias{plotMethodsOverlap}
\title{Plot UpSetR for SummarizedBenchmark object}
\usage{
plotMethodsOverlap(object, assay = "qvalue", alpha = 0.1, ...)
}
\arguments{
\item{object}{A \code{\link{SummarizedBenchmark}} object.}

\item{assay}{The name of an assay.}

\item{alpha}{An alpha value.}

\item{...}{Further arguments passed to \code{\link{upset}}}
}
\value{
An upseR plot.
}
\description{
This function looks for an assay, called by default 'qvalue', and given an alpha threshold,
it binarizes the assay matrix depending on whether its values are below the alpha
threshold. Then it uses the function \code{\link{upset}} to plot the overlaps.
The plot is only generated if at least 2 methods have observations that pass
the alpha threshold.
}
\examples{
data(sb)
\dontrun{
plotMethodsOverlap(sb)
}

}
\seealso{
\code{\link{plotROC}}, \code{\link{estimateMetrics}}
}
\author{
Alejandro Reyes
}
