% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{plotROC}
\alias{plotROC}
\title{Plot ROC curve for SummarizedBenchmark object}
\usage{
plotROC(object, assay = "qvalue")
}
\arguments{
\item{object}{A \code{\link{SummarizedBenchmark}} object.}

\item{assay}{An assay name.}
}
\value{
A ggplot object.
}
\description{
This function inputs a \code{\link{SummarizedBenchmark}} object, looks
for an assay called 'qvalue' and plots receiver operating characteristic curves
for each of the methods to benchmark.
}
\examples{
data(sb)
\dontrun{
plotROC(sb)
}

}
\seealso{
\code{\link{plotMethodsOverlap}}, \code{\link{estimatePerformanceMetrics}}
}
\author{
Alejandro Reyes
}
