% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchDesign.R
\name{print.BenchDesign}
\alias{print.BenchDesign}
\title{Pretty print BenchDesign}
\usage{
\method{print}{BenchDesign}(x, ...)
}
\arguments{
\item{x}{BenchDesign object.}

\item{...}{other parameters.}
}
\description{
Pretty print BenchDesign
}
\author{
Patrick Kimes
}
\keyword{internal}
