% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-soneson2016.R
\docType{data}
\name{sb}
\alias{sb}
\alias{truthdat}
\alias{txi}
\title{SummarizedBenchmark example}
\source{
Example data from the \code{iCOBRA} package.
}
\description{
This object contains the example data from the \code{iCOBRA} package reformatted as a
\code{SummarizedBenchmark} object. It consists of differential expression results from \code{DESeq2}
\code{edgeR} and \code{limma}-voom.
}
\examples{
data(sb)
}
\keyword{SummarizedBenchmark}
\keyword{data}
