% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showBMethod.R
\name{showBMethod}
\alias{showBMethod}
\alias{showBMethods}
\title{Pretty print methods in a BenchDesign}
\usage{
showBMethod(b, n)

showBMethods(b)
}
\arguments{
\item{b}{BenchDesign object.}

\item{n}{name of a method in the BenchDesign to show.}
}
\value{
Brief description is returned to console.
}
\description{
Print out details about a method included in
the BenchDesign. The \code{showBMethods} function is just a wrapper
to call \code{showBMethod} on all methods in the BenchDesign.
}
\examples{
## create empty BenchDesign
bd <- BenchDesign()

## currently no methods
showBMethods(bd)

## add method
bd <- addBMethd(bd, blabel = "method_a", p.adjust)
bd <- addBMethd(bd, blabel = "method_b", qvalue::qvalue)

## show a single method
showBMethod(bd, "method_a")

## show all methods
showBMethods(bd)

}
\author{
Patrick Kimes

Patrick Kimes
}
