% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showMethod.R
\name{showMethod}
\alias{showMethod}
\alias{showMethods}
\title{Pretty print methods in a BenchDesign}
\usage{
showMethod(b, n)

showMethods(b)
}
\arguments{
\item{b}{BenchDesign object}

\item{n}{name of a method in the BenchDesign}
}
\value{
Brief description is returned to console.
}
\description{
Easy method for printing out details about a method included in
the BenchDesign. The \code{showMethods} function is just a wrapper
to call \code{showMethod} on all methods for in the BenchDesign
}
\author{
Patrick Kimes

Patrick Kimes
}
