% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedBenckmark-class.R
\docType{class}
\name{SummarizedBenchmark-class}
\alias{SummarizedBenchmark-class}
\title{SummarizedBenchmark class documentation}
\description{
Extension of the \code{\link{RangedSummarizedExperiment}} to
store the output of different methods intended for the same purpose
in a given dataset. For example, a differential expression analysis could be
done using \pkg{limma}-voom, \pkg{edgeR} and \pkg{DESeq2}. The
SummarizedBenchmark class provides a framework that is useful to store, benckmark and
compare results.
}
\section{Slots}{

\describe{
\item{\code{performanceMetrics}}{A \code{\link{SimpleList}} of the same length
as the number of \code{\link{assays}} containing performance
functions to be compared with the ground truths.}
}}

\author{
Alejandro Reyes
}
