% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDMethod-tidy.R
\docType{methods}
\name{tidyBDMethod}
\alias{tidyBDMethod}
\alias{tidyBDMethod,BDMethod-method}
\title{Tidy BDMethod Data}
\usage{
tidyBDMethod(obj, dat = NULL, ...)

\S4method{tidyBDMethod}{BDMethod}(obj, dat)
}
\arguments{
\item{dat}{optional data object to use when evaluating any
unevaluated expressionsin \code{bdm} meta data.
(default = NULL)xo}

\item{bdm}{a BDMethod object}
}
\value{
A named vector of meta data for the specified BDMethod object, or
a tibble of meta data if a list of BDMethod objects specified.
}
\description{
A helper function to create tabular info for a single
BDMethod object or a list of BDMethod objects.
}
\author{
Patrick Kimes
}
