% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BDMethod-tidy.R
\docType{methods}
\name{tidyBDMethod}
\alias{tidyBDMethod}
\alias{tidyBDMethod}
\alias{tidyBDMethod,BDMethod-method}
\alias{tidyBDMethod,list-method}
\alias{tidyBDMethod,SimpleList-method}
\alias{tidyBDMethod,BenchDesign-method}
\title{Tidy BDMethod Data}
\usage{
tidyBDMethod(obj, dat = NULL, label = FALSE)

\S4method{tidyBDMethod}{BDMethod}(obj, dat)

\S4method{tidyBDMethod}{list}(obj, dat = NULL, label = FALSE)

\S4method{tidyBDMethod}{SimpleList}(obj, dat = NULL, label = FALSE)

\S4method{tidyBDMethod}{BenchDesign}(obj, dat = NULL, label = FALSE)
}
\arguments{
\item{obj}{BDMethod object, list/List of BDMethod objects (e.g. a
BDMethodList), or a BenchDesign object}

\item{dat}{optional data object to use when evaluating any
unevaluated expressionsin \code{bdm} meta data.
(default = NULL)}

\item{label}{logical whether to add a "label" column to the resulting
table containing the names of the methods if \code{obj} was
specified as a named list. (default = FALSE)}
}
\value{
A named vector of meta data if only a single BDMethod object specified, else
a tibble of meta data for the specified list of methods.
}
\description{
A helper function to extract information for a single
or multiple BDMethod object or a list of BDMethod objects.
}
\author{
Patrick Kimes
}
