% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateBench.R
\name{updateBench}
\alias{updateBench}
\title{Check/Update SummarizedBenchmark}
\usage{
updateBench(sb, bd = NULL, dryrun = TRUE, version = FALSE,
  keepAll = TRUE, reuseParams = TRUE, ...)
}
\arguments{
\item{sb}{a \code{\link[=SummarizedBenchmark-class]{SummarizedBenchmark}} object}

\item{bd}{a \code{\link[=BenchDesign-class]{BenchDesign}} object}

\item{dryrun}{logical whether to just print description of what would
be updated rather than actually running any methods. (default = TRUE)}

\item{version}{logical whether to re-run methods with only package
version differences. (default = FALSE)}

\item{keepAll}{logical whether to keep methods run in original \code{\link[=SummarizedBenchmark-class]{SummarizedBenchmark}}
but not in new \code{\link[=BenchDesign-class]{BenchDesign}}. Only used if \code{bd} is not NULL. (default = TRUE)}

\item{reuseParams}{logical whether to reuse parameters from \code{buildBench} call
used to create \code{\link[=SummarizedBenchmark-class]{SummarizedBenchmark}} object (if available). Directly
specified \code{\link{buildBench}} parameters still take precedence. (default = TRUE)}

\item{...}{optional parameters to pass to \code{\link{buildBench}}.}
}
\value{
SumamrizedBenchmark object.
}
\description{
Update \code{\link[=SummarizedBenchmark-class]{SummarizedBenchmark}} results with new methods.
}
\seealso{
\code{\link{buildBench}}
}
\author{
Patrick Kimes
}
