% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateBench.R
\name{updateBench}
\alias{updateBench}
\title{Update SummarizedBenchmark object}
\usage{
updateBench(sb, bd = NULL, dryrun = TRUE, version = FALSE, ...)
}
\arguments{
\item{sb}{a \code{SummarizedBenchmark} object}

\item{bd}{a \code{BenchDesign} object}

\item{dryrun}{logical whether to just print description of what would
be updated rather than actually running any methods. (default = TRUE)}

\item{version}{logical whether to re-run methods with only package
version differences. (default = FALSE)}

\item{...}{optional parameters to pass to \code{\link{buildBench}}.}
}
\value{
SumamrizedBenchmark object.
}
\description{
Update SummarizedBenchmark results with new methods.
}
\author{
Patrick Kimes
}
